Update sp_process.sh
I change the if to see if the last file was a data file, for a if to check if the last expected dat file existed. The same with the *singlepulse.ps file that comes out of the single_pulse_search.py. I also corrected the filename that I pass to SpS, as it was wrong