Add get the information from .hdf5 ^ .singlep file
Provisional notebook to get the information from .hdf5 that are the result of SpS and the files on .singlepulse the result of PRESTO, and the idea is to compare pulses from one and the other. Disclaimer: SpS uses the .singlepulse from PRESTO, is just that I want to check in case it still filters too many pulses that I consider to be of astronomical origin. In general, I wouldn't apply this in other observations that aren't the ones I'm working on right now. So use at your own risk.